Simplify plugin installation
Netbox plugins can now be mounted into the container and will be loaded
automatically from the mounted path. A configuration file to enabled the
plugin in has to be mounted additionally.

To install a plugin from outside the container you need to use the same
Python version as is used in the container. At the moment that is Python 3.11.
Assuming you are in the netbox-docker directory and using the default
configuration, just run this command, to install the netbox-initializers plugin:
pip install --target=plugins netbox-initializers